app/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.1
django-after-response==0.2.2
django-allauth==64.1.0
django-crispy-forms==2.3
django-pwa==2.0.0
gunicorn==23.0.0
packaging==24.1
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
